Get I2cMasterConfig array from TPConfig, fix TP Json parsing, modify TP test json
Add getter function for the array of I2cMasterConfig objects so that the TPConfig
object can be passed to a function "sendTp(...)" in ConfigSender. Modify sendTp
to loop through all I2cMasterConfigs and sends them all. Change test json so that
it contains a few registers in the tp_common_config (including firmware) and a
specific TP with the opc IP and opc node address, which are parsed into global
fields in the TPConfig class through the FEConfig constructor.