[extern] Fix problem with libopen62541 building as shared library
As uncovered by @yzach, `tdaq_cmake` sets `BUILD_SHARED_LIBS` to true.
Internally, `open62541` sets this using a cache variable, which we are
setting `OFF` to build it as a static library to link into
`libnswconfig`.
Explicitly ensuring that `BUILD_SHARED_LIBS` is `OFF` when
`STANDALONE_BUILD_SHARED` is `OFF` resolves the problem that on
reconfiguration of the build files, `libopen62541` is compiled as a
shared library.

Additionally, extra cleanup has been done, more may be needed once it
is figured out why the exports and targets are not available following
the call to `FetchContent_MakeAvailable`.