Disble VMMs for VMM configuration
This commits adds the functionality to disable VMMs (send a config
update to the boards setting register 8 to 0) and enable them again
(set them to the value in the config). This functionality is needed
for the correct configuration procedure where the VMMs are disabled
on the ROC during their configuration. We hope that this removes
the garbage events encountered at the start of run and tests are
looking good so far. The VMMs are disabled after configuring the
ROC, before configuring the VMMs in NSWConfig. They are enabled
again in the prepareForRun transition. The stopRecording transition
disables the VMMs to be symmetric. The nsw::ConfigSender::sendConfig
functionality remains untouched!