[extern] Configurable clone method for FetchContent
This adds a CMake option `NSWCONFIG_FETCH_METHOD`, that will set a
variable `GIT_URL_ENDPOINT` to the appropriate CERN GitLab endpoint,
depending on whether it is set to `ssh`, `https`, `krb`, or `github`.
If nothing is specified, it defaults to `krb`, which seems to be a
sensible default.

It also allows patching of the `open62541-compat` specified `LogIt`
source when building with the bundled `LogIt` (which comes from github
by default).