Move JSON specific API to JsonApi
Allow for integrating OKS device hierarcy with the current JSON config
files as a stopgap until the configuration from Oracle is completed.

`ConfigReader` is fully backwards compatible, but
`ConfigReaderApi` **should not** be used standalone (can't be, as it
is now pure virtual).

Similarly, the `ConfigReaderJsonApi` should not be used standalone
either, for the reason that the interaction should be done through the
`ConfigReader`.

* Base `ConfigReaderApi` is pure virtual
  * Removed `ptree`, `xml`, and OKS APIs
  * `read` is private, `getConfig` returns a `ptree`
* Moved `JsonApi` to its own file
  * Moved most "common" functions to `JsonApi`
  * `JsonApi` has a constructor to take a JSON-like `ptree`
  * `JsonApi::getAllElementNames` returns OKS names in `DeviceMap` if
    it is not empty, otherwise behaves as before

Actions on disabled devices
* Modify ROC ptree if sROC or VMMs are disabled
  * sROC: Disconnect all VMMs from this sROC, disable sROC
  * VMM: Disconnect VMM from sROC and disable VMM
* Skip disabled devices. Needs a partition
* Remove ART, TDS, VMM from ADDC/FEB config tree if disabled

If contains list is empty configure everything in JSON.

Fallback tested in VS setup.