[FetchContent] Try to improve the FetchContent workflow
Enable `UPDATE_DISCONNECTED` for `FetchContent` to prevent recompiling
fetched packages every time CMake reconfigures.

N.B. This change means that if the fetched package is actually updated,
the build will have to be reconfigured from scratch.
Another branch is attempting to optimize this workflow such that it is
transparent to the developer, but it is proving more tricky than expected.